--- a/sklearn/ensemble/tests/test_iforest.py
+++ b/sklearn/ensemble/tests/test_iforest.py
@@ -20,7 +20,7 @@ from sklearn.model_selection import ParameterGrid
 from sklearn.ensemble import IsolationForest
 from sklearn.ensemble._iforest import _average_path_length
 from sklearn.model_selection import train_test_split
-from sklearn.datasets import load_diabetes, load_iris
+from sklearn.datasets import load_diabetes, load_iris, make_classification
 from sklearn.utils import check_random_state
 from sklearn.metrics import roc_auc_score
 
@@ -347,3 +347,12 @@ def test_n_features_deprecation():
 
     with pytest.warns(FutureWarning, match="`n_features_` was deprecated"):
         est.n_features_
+
+
+def test_iforest_with_n_jobs():
+    """This is a non-regression test for PR #23252"""
+    X, _ = make_classification(n_samples=50000, n_features=100)
+    X = csc_matrix(X)
+    X.sort_indices()
+    for n_jobs in [2, 3, 4]:
+        IsolationForest(n_estimators=10, max_samples=256, n_jobs=n_jobs).fit(X)