<!--
Thanks for contributing a pull request! Please ensure you have taken a look at
the contribution guidelines: https://github.com/scikit-learn/scikit-learn/blob/main/CONTRIBUTING.md
-->

#### Reference Issues/PRs
<!--
Example: Fixes #1234. See also #3456.
Please use keywords (e.g., Fixes) to create link to the issues or pull requests
you resolved, so that they will automatically be closed when your pull request
is merged. See https://github.com/blog/1506-closing-issues-via-pull-requests
-->
Fixes #19275
This is a follow-up PR of #23149.

#### What does this implement/fix? Explain your changes.
As discussed in the [comment](https://github.com/scikit-learn/scikit-learn/pull/23149#issuecomment-1107512467), the indexing operation in `_bagging.py: _parallel_build_estimators` is quite expensive, so when `bootstrap_features` is set to **False** and `max_features_` is equal to number of features, we can skip indexing to get better performance.

#### Benchmark result
Code used for profiling:
```python
from sklearn.datasets import make_classification
from scipy.sparse import csc_matrix, csr_matrix
from sklearn.ensemble import IsolationForest

X, y = make_classification(n_samples=50000, n_features=1000)
X = csc_matrix(X)
X.sort_indices()
IsolationForest(n_estimators=10, max_samples=256, n_jobs=1).fit(X)
```
**Sparse Input:**
Before (4.97s)
![image](https://user-images.githubusercontent.com/16646940/164909922-70a89054-a81e-4931-acca-46e4779d8eba.png)


After (0.241s)
![image](https://user-images.githubusercontent.com/16646940/164909946-4c070410-0176-43ed-a94b-cd9e08afb1af.png)


**Dense Input:**
Before (7.99s)
![image](https://user-images.githubusercontent.com/16646940/164909872-3cec6573-efd9-4baf-83a2-d47caaec7e14.png)

After (5.19s)
![image](https://user-images.githubusercontent.com/16646940/164909792-52b711b6-0280-44cc-8566-036ba0043162.png)


#### Any other comments?
`boostrap_features` is set to **False** already for IsolationForest, but users can still set `max_features` freely, which will cause a worse runtime when it's not 1.0. 

Since in IsolationForest, we're randomly picking a single feature to split, shall we deprecate the `max_features` argument and always set it to 1.0 for better runtime ?

<!--
Please be aware that we are a loose team of volunteers so patience is
necessary; assistance handling other issues is very welcome. We value
all user contributions, no matter how minor they are. If we are slow to
review, either the pull request needs some benchmarking, tinkering,
convincing, etc. or more likely the reviewers are simply busy. In either
case, we ask for your understanding during the review process.
For more information, see our FAQ on this topic:
http://scikit-learn.org/dev/faq.html#why-is-my-pull-request-not-getting-any-attention.

Thanks for contributing!
-->