@MaxwellLZH 2022-05-01
> Can you update the opening comment to include the benchmark script so it is easier for other reviewers to run the benchmark? (without needing to find it in the other PR)

Sure! I've updated the benchmark result in the opening comment.