@ogrisel 2022-05-02
I get a segfault when I execute https://github.com/scikit-learn/scikit-learn/issues/19275#issuecomment-768166443 with this branch.

Here is the top of the gdb backtrace:


```
#0  0x00007fffb4b2b6db in __pyx_f_7sklearn_4tree_9_splitter_18BaseSparseSplitter_extract_nnz () from /home/ogrisel/code/scikit-learn/sklearn/tree/_splitter.cpython-310-x86_64-linux-gnu.so
#1  0x00007fffb4b33722 in __pyx_f_7sklearn_4tree_9_splitter_20RandomSparseSplitter_node_split () from /home/ogrisel/code/scikit-learn/sklearn/tree/_splitter.cpython-310-x86_64-linux-gnu.so
#2  0x00007fffb4b01e4a in __pyx_f_7sklearn_4tree_5_tree_21DepthFirstTreeBuilder_build(__pyx_obj_7sklearn_4tree_5_tree_DepthFirstTreeBuilder*, __pyx_obj_7sklearn_4tree_5_tree_Tree*, _object*, tagPyArrayObject_fields*, int, __pyx_opt_args_7sklearn_4tree_5_tree_21DepthFirstTreeBuilder_build*) () from /home/ogrisel/code/scikit-learn/sklearn/tree/_tree.cpython-310-x86_64-linux-gnu.so
#3  0x00007fffb4ac4970 in __pyx_pw_7sklearn_4tree_5_tree_21DepthFirstTreeBuilder_3build(_object*, _object*, _object*) () from /home/ogrisel/code/scikit-learn/sklearn/tree/_tree.cpython-310-x86_64-linux-gnu.so
#4  0x00005555556997f6 in method_vectorcall_VARARGS_KEYWORDS (func=<optimized out>, args=<optimized out>, nargsf=<optimized out>, kwnames=<optimized out>)
    at /usr/local/src/conda/python-3.10.4/Objects/descrobject.c:344
#5  0x0000555555687f2f in _PyObject_VectorcallTstate (kwnames=0x0, nargsf=<optimized out>, args=0x7fffa4001c40, callable=0x7fffb4c64400, tstate=0x55555673f3d0)
    at /usr/local/src/conda/python-3.10.4/Include/cpython/abstract.h:114
#6  PyObject_Vectorcall (kwnames=0x0, nargsf=<optimized out>, args=0x7fffa4001c40, callable=0x7fffb4c64400) at /usr/local/src/conda/python-3.10.4/Include/cpython/abstract.h:123
#7  call_function (kwnames=0x0, oparg=<optimized out>, pp_stack=<synthetic pointer>, trace_info=0x7fffb2ba1560, tstate=<optimized out>) at /usr/local/src/conda/python-3.10.4/Python/ceval.c:5867
#8  _PyEval_EvalFrameDefault (tstate=<optimized out>, f=<optimized out>, throwflag=<optimized out>) at /usr/local/src/conda/python-3.10.4/Python/ceval.c:4198
#9  0x00005555556a56ac in _PyEval_EvalFrame (throwflag=0, f=0x7fffa4001a10, tstate=0x55555673f3d0) at /usr/local/src/conda/python-3.10.4/Include/internal/pycore_ceval.h:46
#10 _PyEval_Vector (kwnames=<optimized out>, argcount=<optimized out>, args=0x7fffb3afeed0, locals=0x0, con=0x7fffb4c46690, tstate=0x55555673f3d0) at /usr/local/src/conda/python-3.10.4/Python/ceval.c:5065
#11 _PyFunction_Vectorcall (kwnames=<optimized out>, nargsf=<optimized out>, stack=0x7fffb3afeed0, func=0x7fffb4c46680) at /usr/local/src/conda/python-3.10.4/Objects/call.c:342
#12 _PyObject_VectorcallTstate (kwnames=<optimized out>, nargsf=<optimized out>, args=0x7fffb3afeed0, callable=0x7fffb4c46680, tstate=0x55555673f3d0)
    at /usr/local/src/conda/python-3.10.4/Include/cpython/abstract.h:114
```