@MaxwellLZH 2022-05-02
I tried different settings, it seems like the segmentation fault only happens when using sparse data and `n_jobs > 1`.