@ogrisel 2022-05-02
There are several places in the code where the data is mutated in place. This is not a problem when working with `X[:, features]` because this forces a copy each time. But this becomes a problem in your branch because then `X` is shared across concurrent thread and therefore the tree-splitting code is no longer thread-safe.

We need to sort out how to make this thread-safe without introducing costly data-copies.