@ogrisel 2022-05-02
I confirm that making an explicit copy of `X` fixes the problem:

```diff
diff --git a/sklearn/ensemble/_bagging.py b/sklearn/ensemble/_bagging.py
index 9c8faa783..7798dc880 100644
--- a/sklearn/ensemble/_bagging.py
+++ b/sklearn/ensemble/_bagging.py
@@ -137,10 +137,10 @@ def _parallel_build_estimators(
                 not_indices_mask = ~indices_to_mask(indices, n_samples)
                 curr_sample_weight[not_indices_mask] = 0
 
-            X_ = X[:, features] if requires_feature_indexing else X
+            X_ = X[:, features] if requires_feature_indexing else X.copy()
             estimator_fit(X_, y, sample_weight=curr_sample_weight)
         else:
-            X_ = X[indices][:, features] if requires_feature_indexing else X[indices]
+            X_ = X[indices][:, features] if requires_feature_indexing else X[indices].copy()
             estimator_fit(X_, y[indices])
 
         estimators.append(estimator)
```

However, I still do not understand why it's needed: there should be no inplace modification of X.