@thomasjpfan 2022-05-02
With `check_input=False`, the tree code does not get to run:

https://github.com/scikit-learn/scikit-learn/blob/452ede02c681f28ef5b414e73456ba6bd783a2fe/sklearn/tree/_classes.py#L170

so it will do an inplace operation here that changes the dtype of the data:

https://github.com/scikit-learn/scikit-learn/blob/452ede02c681f28ef5b414e73456ba6bd783a2fe/sklearn/tree/_tree.pyx#L102-L103

The quick fix is:

```diff
diff --git a/sklearn/ensemble/_iforest.py b/sklearn/ensemble/_iforest.py
index 4be74d2873..07d28c93e2 100644
--- a/sklearn/ensemble/_iforest.py
+++ b/sklearn/ensemble/_iforest.py
@@ -254,7 +254,7 @@ class IsolationForest(OutlierMixin, BaseBagging):
         self : object
             Fitted estimator.
         """
-        X = self._validate_data(X, accept_sparse=["csc"])
+        X = self._validate_data(X, accept_sparse=["csc"], dtype=np.float32)
         if issparse(X):
             # Pre-sort indices to avoid that each individual tree of the
             # ensemble sorts the indices.
```