@MaxwellLZH 2022-05-03
I've update the code as suggested by @thomasjpfan , and verifies that it fixes the segmentation fault issue :)