@ogrisel 2022-05-03
Could you please add a non-regression test?