@ogrisel 2022-05-03
Please also add a new changelog entry targeting scikit-learn 1.2 (unless our 1.1 release manager @jeremiedbb thinks it should be part of 1.1 :P)