@MaxwellLZH 2022-05-03
Hi @ogrisel , I've added test case and changelog as suggested, meanwhile I think we still need to make a decision on whether we should deprecate the `max_features` argument and set it to 1.0 at all times before merging this PR. 

Currently users may set `max_features` to a smaller value and observe a much slower runtime, which could be quite confusing :(