@ogrisel 2022-05-03
I think we can keep `max_features`. Maybe we can just add a note to explain that it's much slower to enable feature subsampling on sparse training data for this model.